moved all data classes inside one kotlin file
